+++ preprocessing/nextclade/src/loculus_preprocessing/processing_functions.py
@@ -25,7 +25,6 @@ from urllib3.util.retry import Retry
 from .datatypes import (
     AnnotationSource,
     AnnotationSourceType,
-    FileIdAndName,
     FunctionArgs,
     InputData,
     InputMetadata,
@@ -1947,39 +1946,6 @@ def single_metadata_annotation(
             message=message,
         )
     ]
-
-
-def validate_raw_reads_submission(
-    files: list[FileIdAndName],
-) -> tuple[list[ProcessingAnnotation], list[ProcessingAnnotation]]:
-    errors: list[ProcessingAnnotation] = []
-    warnings: list[ProcessingAnnotation] = []
-
-    if len(files) > 2:  # noqa: PLR2004
-        message = f"Raw reads must be submitted as one or two files, got {len(files)}"
-        errors.append(
-            ProcessingAnnotation.from_fields(
-                input_fields=[f.name for f in files],
-                output_fields=[f.name for f in files],
-                type=AnnotationSourceType.SUBMITTED_FILE,
-                message=message,
-            )
-        )
-
-    allowed_extensions = [".fastq", ".fastq.gz", ".fq", ".fq.gz"]
-    for file in files:
-        if not any(file.name.endswith(extension) for extension in allowed_extensions):
-            message = (
-                f"Raw reads file '{file.name}' has unrecognized extension."
-                f" Allowed extensions: {', '.join(allowed_extensions)}"
-            )
-            errors.append(
-                ProcessingAnnotation.from_single(
-                    name=file.name, type=AnnotationSourceType.SUBMITTED_FILE, message=message
-                )
-            )
-
-    return errors, warnings
 
 
 def process_frameshifts(input: str | None) -> InputData:
